+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes will be documented in this file.
 
 ## Head
 
+* Support PAPI modify order (#487)
 * `ExchangeInfo.contractStatus` was not being processed for COIN-M products (#485)
 
 ## 1.0.4 &ndash; 2024-12-30